ocr_eval/requirements.txt
aniso8601==9.0.1
attrs==23.1.0
Babel==2.9.1
blinker==1.6.2
cffi==1.14.5
chardet==4.0.0
clams-python==1.0.9
click==8.1.3
cmake==3.27.2
cryptography==36.0.1
dasbus==1.4
dbus-python==1.2.18
decorator==4.4.2
deepdiff==6.3.0
dnspython==2.1.0
ethtool==0.15
fail2ban==1.0.2
ffmpeg-python==0.2.0
file-magic==0.4.0
filelock==3.12.2
Flask==2.3.2
Flask-RESTful==0.3.10
future==0.18.3
gpg==1.15.1
gssapi==1.6.9
gunicorn==20.1.0
idna==2.10
importlib-metadata==6.6.0
iniparse==0.4
ipaclient==4.10.0
ipalib==4.10.0
ipaplatform==4.10.0
ipapython==4.10.0
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
jwcrypto==0.8
lapps==0.0.2
libcomps==0.1.18
lightning-utilities==0.9.0
lit==16.0.6
lxml==4.6.5
MarkupSafe==2.1.3
mmif-docloc-baapb==230817
mmif-python==1.0.8
mpmath==1.3.0
netaddr==0.8.0
netifaces==0.10.6
networkx==3.1
nftables==0.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
ordered-set==4.1.0
packaging==23.1
pandas==2.0.3
perf==0.1
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.5.3
ply==3.11
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.10.9
PyGObject==3.40.1
pyinotify==0.9.6
pyrsistent==0.19.3
PySocks==1.7.1
python-augeas==0.5.0
python-dateutil==2.8.2
python-ldap==3.3.1
python-linux-procfs==0.7.0
python-yubico==1.3.3
pytz==2021.1
pyudev==0.22.0
pyusb==1.0.2
PyYAML==5.4.1
qrcode==6.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==1.0.0
rpm==4.16.1.3
selinux==3.4
sepolicy==3.4
setools==4.4.0
setroubleshoot==1.1
six==1.15.0
sos==4.5.1
SSSDConfig==2.7.3
subscription-manager==1.29.30.1
sympy==1.12
systemd-python==234
torch==2.0.1
torchmetrics==1.0.3
triton==2.0.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.5
Werkzeug==2.3.6
zeep==4.2.1
zipp==3.15.0
